Extract IrGraphGenerator
